+++ apps/web/app/api/workspaces/[idOrSlug]/billing/invoices/route.ts
@@ -67,10 +67,7 @@ const payoutInvoices = async (workspaceId: string) => {
     return {
       ...invoice,
       description: "Dub Partner payout",
-      pdfUrl:
-        invoice.status === "completed"
-          ? `${APP_DOMAIN}/api/workspaces/ws_${workspaceId}/billing/invoices/${invoice.id}`
-          : null,
+      pdfUrl: `${APP_DOMAIN}/invoices/${invoice.id}`,
     };
   });
 };
