+++ apps/web/app/api/workspaces/route.ts
@@ -6,7 +6,11 @@ import {
   createWorkspaceSchema,
 } from "@/lib/zod/schemas/workspaces";
 import { prisma } from "@dub/prisma";
-import { FREE_WORKSPACES_LIMIT, nanoid } from "@dub/utils";
+import {
+  FREE_WORKSPACES_LIMIT,
+  generateRandomString,
+  nanoid,
+} from "@dub/utils";
 import { waitUntil } from "@vercel/functions";
 import { NextResponse } from "next/server";
 
@@ -93,6 +97,7 @@ export const POST = withSession(async ({ req, session }) => {
           },
         },
         billingCycleStart: new Date().getDate(),
+        invoicePrefix: generateRandomString(8),
         inviteCode: nanoid(24),
         defaultDomains: {
           create: {}, // by default, we give users all the default domains when they create a project
